% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{data}
\name{brfss}
\alias{brfss}
\title{Behavioral Risk Factor Surveillance System 2017}
\format{A data frame with 444023 observations on the following 10 variables.
\describe{
\item{\code{OBESE}}{(original : \code{_RFBMI5}) Adults who have a body mass index greater than 25.00 (Overweight or Obese)

(1) No \cr (2) Yes}
\item{\code{PA300}}{(original : \code{_PA300R2}) Adults that participated in 300 minutes (or vigorous equivalent minutes) of physical activity per week.

(1) 301+ minutes \cr (2) 1-300 minutes \cr (3) 0 minutes
\cr
}
\item{\code{FRTLT1A}}{(original : \code{_FRTLT1A}) Consume Fruit 1 or more times per day

(1) Consumed fruit one or more times per day \cr
(2) Consumed fruit less than one time per day}
\item{\code{VEGLT1A}}{(original : \code{_VEGLT1A}) Consume Vegetables 1 or more times per day

(1) Consumed vegetables one or more times per day \cr
(2) Consumed vegetables less than one time per day}
\item{\code{SMOKER}}{(original : \code{_SMOKER3}) Four-level smoker status: Everyday smoker, Someday smoker, Former smoker, Non-smoker

(1) Current smoker - now smokes every day\cr
(2) Current smoker - now smokes some days\cr
(3) Former smoker\cr
(4) Never smoked
}
\item{\code{DRNK30}}{(original : \code{DRNKANY5}) Adults who reported having had at least one drink of alcohol in the past 30 days.

(1) Yes\cr
(2) No}
\item{\code{SEX}}{Respondents Sex}
\item{\code{INCOME}}{(original : \code{INCOME2}) Income Level}
\item{\code{STATE}}{(original : \code{_STATE}) State FIPS Code}
\item{\code{REGION}}{Region to which the states belongs}
\item{\code{PARTY}}{The party that won the 2016 presidential election by state}
}}
\source{
\url{https://www.cdc.gov/brfss/}

\url{https://doi.org/10.7910/DVN/42MVDX}
}
\description{
This data is comprised of questions related to healthy lifestyles, such as obesity, exercise time, eating habits, smoking and drinking, among the 2017 BRFSS survey, and includes covariates such as respondents’ gender and income levels. There are also states in which the respondent resides as a group variable, regions in which the state belongs, and parties in which the state won the 2016 presidential election.
}
\examples{
data("brfss")
brfss2 = brfss[sample(1:nrow(brfss), 2000),]
lca = glca(item(OBESE, PA300, FRTLT1A, VEGLT1A, SMOKER, DRNK30) ~ 1,
   data = brfss2, nclass = 3)
summary(lca)

lcr = glca(item(OBESE, PA300, FRTLT1A, VEGLT1A, SMOKER, DRNK30) ~ SEX,
   group = INCOME, data = brfss2, nclass = 3)
summary(lcr)
coef(lcr)

mglca = glca(item(OBESE, PA300, FRTLT1A, VEGLT1A, SMOKER, DRNK30) ~ 1,
   group = SEX, data = brfss2, nclass = 3)
summary(mglca)

mlca = glca(item(OBESE, PA300, FRTLT1A, VEGLT1A, SMOKER, DRNK30) ~ 1,
   group = STATE, data = brfss2, nclass = 3, ncluster = 2)
summary(mlca)

mlcr = glca(item(OBESE, PA300, FRTLT1A, VEGLT1A, SMOKER, DRNK30) ~ SEX,
   group = STATE, data = brfss2, nclass = 3, ncluster = 2)
summary(mlcr)
coef(mlcr)

mlcr2 = glca(item(OBESE, PA300, FRTLT1A, VEGLT1A, SMOKER, DRNK30) ~ SEX + PARTY,
   group = STATE, data = brfss2, nclass = 3, ncluster = 2)
summary(mlcr2)
coef(mlcr2)
}
\references{
Centers for Disease Control and Prevention (CDC). Behavioral Risk Factor Surveillance System Survey Data. Atlanta, Georgia: U.S. Department of Health and Human Services, Centers for Disease Control and Prevention, [2017].

MIT Election Data and Science Lab, 2017, "U.S. President 1976–2016", Harvard Dataverse, V5.
}
\keyword{datasets}
