% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.glca.R
\name{coef.glca}
\alias{coef.glca}
\title{Extract \code{glca} Model Coefficients.}
\usage{
\method{coef}{glca}(object, digits = max(3, getOption("digits") - 3),
  show.signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{object}{an object of "\code{glca}", usually, a result of a call to \code{glca}.}

\item{digits}{the number of significant digits to use when printing.}

\item{show.signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Coefficient matrix from the \code{glca} model.

If the model has been calculated standard error, coefficient matrix contains standard error and t, p-value.
}
\description{
Basically, show AIC and BIC for model fit criteria. When \code{nboot} given, using parametric bootstrap, compute bootstrap p-value for "relative" and "absolute" model fit.
}
\examples{
## For examples see example(glca)

}
\seealso{
\code{\link{glca}}
}
