% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{glca-package}
\alias{glca-package}
\title{Latent Class Analysis with Grouped Data}
\description{
Fitting latent class models for data containing group variable. \cr Standard LCA(LCA), Multiple-Group LCA(MGLCA), Multilevel LCA(MLCA) can be fitted.
}
