% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glca.gof.R
\name{glca.gof}
\alias{glca.gof}
\title{Goodness of Fit Tests for Fitted \code{glca} Model (AIC, BIC, Bootstrap p-value)}
\usage{
glca.gof(object, object2 = NULL, nboot = 25, random.seed = NULL,
  maxiter = 1000, eps = 1e-06, verbose = TRUE)
}
\arguments{
\item{object}{an object of "\code{glca}", usually, a result of a call to \code{glca}.}

\item{object2}{an optional object of "\code{glca}" to be compared with former object.}

\item{nboot}{number of bootstrap samples.}

\item{random.seed}{random seed which generates the seed from \code{R}}

\item{maxiter}{an integer for maximum number of iteration for bootstrap sample.}

\item{eps}{positive convergence tolerance for bootstrap sample.}

\item{verbose}{an logical value for whether or not to print the result of a function's execution.}
}
\value{
\item{table}{a table with model fit criteria and bootstrap p-value.}
\item{boot}{a list of \eqn{G^2} statistics from each bootstrap sample}
}
\description{
Provides model fit criteria such as AIC, BIC, deviance statistics\eqn{(G^2)} and bootstrap p-value for "absolute" model fit. Given \code{object2}, it also computes bootstrap p-value for "relative" model fit. The bootstrap p-value is obtained from the empirical distribution of \eqn{G^2} statistics.
}
\examples{
## Example 1.
## LCA model selection between two models with different number of latent classes.
data(gss)
class4 = glca(item(ABDEFECT, ABNOMORE, ABHLTH, ABPOOR, ABRAPE, ABSINGLE) ~ 1,
              data = gss, n_init = 10, na.rm = TRUE, nclass = 4)
class5 = glca(item(ABDEFECT, ABNOMORE, ABHLTH, ABPOOR, ABRAPE, ABSINGLE) ~ 1,
              data = gss, n_init = 10, na.rm = TRUE, nclass = 5)

\dontrun{glca.gof(class4, class5)}

## Example 2.
## MGLCA model selection between two models with different number of latent clusters.
cluster3 = glca(item(ABDEFECT, ABNOMORE, ABHLTH, ABPOOR, ABRAPE, ABSINGLE) ~ 1,
                group = REGION, data = gss, nclass = 3, ncluster = 2)
cluster4 = glca(item(ABDEFECT, ABNOMORE, ABHLTH, ABPOOR, ABRAPE, ABSINGLE) ~ 1,
                group = REGION, data = gss, nclass = 3, ncluster = 3)

\dontrun{glca.gof(cluster3, cluster4)}

## Example 3.
## MGLCA model selection under the measurement (invariance) assumption across groups.
measInv = glca(item(ABDEFECT, ABNOMORE, ABHLTH, ABPOOR, ABRAPE, ABSINGLE) ~ 1,
               group = SEX, n_init = 10, data = gss, nclass = 4)
measVar = glca(item(ABDEFECT, ABNOMORE, ABHLTH, ABPOOR, ABRAPE, ABSINGLE) ~ 1,
               group = SEX, n_init = 10, data = gss, nclass = 4, measure_inv = FALSE)

\dontrun{glca.gof(measInv, measVar)}

}
\references{
Akaike H (1974). \emph{A New Look at the Statistical Model Identification.} IEEE Transactions on Automatic Control, 19, 716–723. \url{http://doi.org/10.1109/tac.1974.1100705}

Schwarz G (1978). \emph{Estimating the Dimensions of a Model.} The Annals of Statistics, 6, 461–464. \url{http://doi.org/10.1214/aos/1176344136}

LANGEHEINE, ROLF & Pannekoek, Jeroen & van de Pol, Frank. (1996). \emph{Bootstrapping Goodness-of-Fit Measures in Categorical Data Analysis.} Sociological Methods &amp Research. 24. 492-516. \url{http://doi.org/10.1177/0049124196024004004}
}
\seealso{
\code{\link{glca}} \code{\link{gss}}
}
\author{
Youngsun Kim
}
