% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\name{item}
\alias{item}
\title{Specify Manifest Items for \code{glca}.}
\usage{
item(...)
}
\arguments{
\item{\dots}{vectors of manifest items. These can be given as named arguments which is colnames of data.frame}
}
\value{
a \code{matrix} of specified variables, which contains following attributes about items.

Dimensions of \code{data.frmae}, names of items and level of items.
}
\description{
Specifying manifest items in formula of \code{glca} function.
}
\examples{
## For examples see example(glca)

}
\seealso{
\code{\link{glca}}
}
