% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{National Health and Nutrition Examination Survey 2015-2016}
\format{A data frame with 5735 observations on the following 8 variables.
\describe{
\item{\code{DPQ010}}{Have little interest in doing things}
\item{\code{DPQ020}}{Feeling down, depressed, or hopeless}
\item{\code{DPQ030}}{Trouble sleeping or sleeping too much}
\item{\code{DPQ040}}{Feeling tired or having little energy}
\item{\code{DPQ050}}{Poor appetite or overeating}
\item{\code{AGE}}{Age in years at screening}
\item{\code{GENDER}}{Respondent's Gender}
\item{\code{RACE}}{Race/Hispanic origin w/ NH Asian}
}}
\source{
\url{https://wwwn.cdc.gov/Nchs/Nhanes/}
}
\description{
There is a nine-item depression screening instrument of National Health and Nutrition Examination Survey, which can indicating mental health of respondents. Among thoes items, former 5 items are selected. And the covariates include gender, age, and race of respondents.
}
\examples{
data("nhanes")

# Model 1: LCA
lca = glca(item(DPQ010, DPQ020, DPQ030, DPQ040, DPQ050) ~ 1,
           data = nhanes, nclass = 2)
summary(lca)

# Model 2: LCA with covariate(s)
lcr = glca(item(DPQ010, DPQ020, DPQ030, DPQ040, DPQ050) ~ AGE,
           data = nhanes, nclass = 3)
summary(lcr)
coef(lcr)

# Model 3: MGLCA
mglca = glca(item(DPQ010, DPQ020, DPQ030, DPQ040, DPQ050) ~ 1,
             group = GENDER, data = nhanes, nclass = 3)
summary(mglca)

# Model 4: MGLCA with covariate(s)
mglcr = glca(item(DPQ010, DPQ020, DPQ030, DPQ040, DPQ050) ~ AGE,
             group = GENDER, data = nhanes, nclass = 3)
summary(mglcr)
coef(mglcr)

# Model 5: MLCA
mlca = glca(item(DPQ010, DPQ020, DPQ030, DPQ040, DPQ050) ~ 1,
            group = RACE, data = nhanes, nclass = 3, ncluster = 2)
summary(mlca)
}
\references{
Centers for Disease Control and Prevention (CDC). National Center for Health Statistics (NCHS). National Health and Nutrition Examination Survey Data. Hyattsville, MD: U.S. Department of Health and Human Services, Centers for Disease Control and Prevention, [2015-2016][\url{https://wwwn.cdc.gov/nchs/nhanes/ContinuousNhanes/Default.aspx?BeginYear=2015}]
}
\keyword{datasets}
