% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{data}
\name{nyts18}
\alias{nyts18}
\title{National Youth Tobacco Survey (NYTS) 2018}
\format{
A data frame with 3018 observations on the following 8 variables.
\describe{
\item{\code{CIGRT}}{}
\item{\code{CIGAR}}{}
\item{\code{CHEW}}{}
\item{\code{E_CIGAR}}{}
\item{\code{HOOKAH}}{}

\item{\code{SEX}}{Respondents Sex}
\item{\code{Sch_ID}}{}
\item{\code{Sch_type}}{}
}
}
\source{
\url{https://www.cdc.gov/tobacco/data_statistics/surveys/nyts/index.htm}
}
\description{
National Youth Tobacco Survey (NYTS) 2018
}
\examples{
data("nyts18")
\donttest{
# Model 1: LCA
lca = glca(item(CIGRT, CIGAR, CHEW, E_CIGAR, HOOKAH) ~ 1,
           data = nyts18, nclass = 3)
summary(lca)

# Model 2: LCR
lca = glca(item(CIGRT, CIGAR, CHEW, E_CIGAR, HOOKAH) ~ SEX,
           data = nyts18, nclass = 3)
summary(lca)
coef(lca)

# Model 3: MGLCA
mglca = glca(item(CIGRT, CIGAR, CHEW, E_CIGAR, HOOKAH) ~ 1,
             group = SEX, data = nyts18, nclass = 3)
summary(mglca)

# Model 4: MLCA
mlca = glca(item(CIGRT, CIGAR, CHEW, E_CIGAR, HOOKAH) ~ 1,
   group = Sch_ID, data = nyts18, nclass = 3, ncluster = 2)
summary(mlca)

# Model 5: MLCA with level-1 covariate(s) only
mlcr = glca(item(CIGRT, CIGAR, CHEW, E_CIGAR, HOOKAH) ~ SEX,
            group = Sch_ID, data = nyts18, nclass = 3, ncluster = 2)
summary(mlcr)
coef(mlcr)

# Model 6: MLCA with level-1 and level-2 covariate(s)
# (SEX: level-1 covariate, PARTY: level-2 covariate)
mlcr2 = glca(item(CIGRT, CIGAR, CHEW, E_CIGAR, HOOKAH) ~ SEX + Sch_type,
             group = Sch_ID, data = nyts18, nclass = 3, ncluster = 2)
summary(mlcr2)
coef(mlcr2)
}
}
\keyword{datasets}
