\name{summary.glca}
\alias{summary.glca}
\title{
Summarize the Estimated Parameter of fitted gLCA obejct.
}
\description{
\code{summary} method for class "\code{glca}".
}
\usage{
## S3 method for class 'glca'
summary(object, digits = max(3, getOption("digits") - 3))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{an object of "\code{glca}", usually, a result of a call to \code{glca}.}
  \item{digits}{the number of digits to be printed.}
}
\details{

}
\value{
This function prints decriptions of model and its more detailed estimated parameters but returns \code{NULL}.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Youngsun Kim
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{glca}}
}
\examples{
## For examples see example(glca)
}
