% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glca.R
\name{glca}
\alias{glca}
\title{Fits Latent Class Analysis Including Group Variable and Covariates}
\usage{
glca(formula, group = NULL, data, nclass = 3, ncluster = 0,
  measure_inv = TRUE, std_err = TRUE, init_param = NULL,
  n_init = 1, maxiter = 5000, eps = 1e-10, na.rm = FALSE,
  random.seed = NULL, verbose = TRUE)
}
\arguments{
\item{formula}{a formula for specifying manifest items using the "\code{item}" function and covariates.}

\item{group}{an optional vector specifying a group of observations, which can be include group covariates using "\code{group}" function. given group variable, group covariates can be incorporated.}

\item{data}{a data frame containing the manifest item, covariates and group variable.}

\item{nclass}{number of latent classes. default is 3.}

\item{ncluster}{number of level 2 latent classes. given ncluster and group variable, the MLCA will be fitted.}

\item{measure_inv}{a logical value of the assumption of measurement invariance across groups.}

\item{std_err}{a logical value whether calculating standard error of estimates. default is TRUE.}

\item{init_param}{a list which contains user-defined initial parameter.}

\item{n_init}{number of random initial parameter set.}

\item{maxiter}{an integer for maximum number of iteration.}

\item{eps}{positive convergence tolerance.}

\item{na.rm}{a logical value whether or not to remove observations missing at least 1 item.}

\item{random.seed}{random seed which generates the seed from R}

\item{verbose}{a logical value for whether or not to print the result of a function's execution.}
}
\value{
\code{glca} returns an object of class "\code{glca}".

The function \code{summary} prints estimates for parameters and \code{glca.gof} function gives goodness of fit measures for the model.

An object of class "\code{glca}" is a list containing at least the following components:

\item{call}{the matched call.}
\item{model}{a list which contains model descriptions.}
\item{datalist}{a list of data used for fitting.}
\item{param}{a list of parameter estimates.}
\item{std.err}{a list of standard error for estimates}
\item{coefficient}{a list of model coefficients for prevalence.}
\item{posterior}{a data frame with posterior probablity of each individaul for latent classes and each group for latent clusters}
\item{count}{a data frame with unique patterns of manifest items and corresponding observed and predicted counts.}
\item{gof}{a list of goodness of fit measures, i.e. AIC, BIC, and log-likelihood.}
\item{convergence}{a list about convergence which contains whether or not it has been converged, number of iterations and scores.}
}
\description{
Function for fitting LCA models with multiple groups, which may include latent class structure for group variable.
}
\details{
The formula should consist of an \code{~} operator between two sides. Manifest items should be indicated in LHS of formula using \code{item} function and covariates should be specified in RHS of formula. For example, \cr
\code{item(y1, y2, y3) ~ 1}\cr
\code{item(y1, y2, y3) ~ x1 + x2}

In \code{glca}, manifest items should be categorical variable and the probability of each category of manifest items varies across latent classes. This probabilities are called "item response probability" (\code{rho} in \code{glca} output). The latent variables (i.e latent class or latent cluster) are categorical as well. The parameters indicating probability of each category is called "prevalence" (\code{delta}, \code{gamma} in \code{glca} output). The prevalence for latent class (\code{gamma}) can be modeled with covariates. Using logistic regression, coefficients for each covariates (\code{beta} in \code{glca} output) can be estimated.

\code{glca} can handle two types of covariates. If covariates vary across individuals, the covariates are considered as level-1 covariates. Given group, and the covariates varying across groups are considered as level-2 covariates. Both type of covariates have an effect on class prevalence. These covariate type will be automatically detected by \code{glca}.
}
\examples{
##
## Example 1. GSS dataset
##
data("gss")

# LCA
lca = glca(item(ABDEFECT, ABNOMORE, ABHLTH, ABPOOR, ABRAPE, ABSINGLE) ~ 1,
            data = gss, nclass = 4)
summary(lca)

# LCA with covariate(s)
lcr = glca(item(ABDEFECT, ABNOMORE, ABHLTH, ABPOOR, ABRAPE, ABSINGLE) ~ AGE,
           data = gss, nclass = 4)
summary(lcr)
coef(lcr)

# Multitple Group LCA (MGLCA)
mglca = glca(item(ABDEFECT, ABNOMORE, ABHLTH, ABPOOR, ABRAPE, ABSINGLE) ~ 1,
             group = DEGREE, data = gss, nclass = 4)
summary(mglca)

# Multitple Group LCA with Covariates (MGLCR)
mglcr = glca(item(ABDEFECT, ABNOMORE, ABHLTH, ABPOOR, ABRAPE, ABSINGLE) ~ SEX,
             group = DEGREE, data = gss, nclass = 3)
summary(mglcr)
coef(mglcr)

##
## Example 2. BRFSS data
##
data("brfss")

# Multilevel LCA (MLCA)
brfss1000 = brfss[sample(1:nrow(brfss), 1000),]
mlca = glca(item(OBESE, PA300, FRTLT1A, VEGLT1A, SMOKER, DRNK30) ~ 1,
            group = STATE, data = brfss1000, nclass = 3, ncluster = 2)
summary(mlca)

# MLCA with Covariates (MLCR)
# (SEX: level-1 covariate, REGION: level-2 covariate)
mlcr = glca(item(OBESE, PA300, FRTLT1A, VEGLT1A, SMOKER, DRNK30) ~ SEX + REGION,
            group = STATE, data = brfss1000, nclass = 3, ncluster = 2)
summary(mlcr)
coef(mlcr)

}
\references{
Jeroen K. Vermunt (2003). \emph{7. Multilevel Latent Class Models}. Sociological Methodology, 33(1), 213–239. \url{https://doi.org/10.1111/j.0081-1750.2003.t01-1-00131.x}

Linda M. Collins, Stephanie T. Lanza (2009). \emph{Latent Class and Latent Transition Analysis: With Applications in the Social, Behavioral, and Health Sciences}. John Wiley & Sons Inc.
}
\seealso{
\code{\link{gss}} \code{\link{brfss}}
}
\author{
Youngsun Kim
}
