% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\name{item}
\alias{item}
\title{Specifies Manifest Items for \code{glca}}
\usage{
item(..., starts.with = NULL, ends.with = NULL)
}
\arguments{
\item{\dots}{vectors of manifest items. These can be given as named arguments which is colnames of \code{data.frame}.}

\item{starts.with}{a string for prefix of variable names to be selected.}

\item{ends.with}{a string for suffix of variable names to be selected.}
}
\value{
a \code{matrix} of specified variables, which contains names and levels of manifest items.
}
\description{
Specifying manifest items in formula of \code{glca} function.
}
\examples{
## For examples see example(glca)

}
\seealso{
\code{\link{glca}}
}
