% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.glca.R
\name{plot.glca}
\alias{plot.glca}
\title{Summarize the Estimated Parameters of Fitted glca Model}
\usage{
\method{plot}{glca}(x, group.name = NULL, ...)
}
\arguments{
\item{x}{an object of "\code{glca}", usually, a result of a call to \code{glca}}

\item{...}{further arguments passed to or from other methods}
}
\value{
This function prints decriptions of model and its more detailed estimated parameters but returns \code{NULL}.
}
\description{
\code{summary} method for class "\code{glca}".
}
\examples{
## For examples see example(glca)

}
\seealso{
\code{\link{glca}}
}
