% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.glca.R
\name{summary.glca}
\alias{summary.glca}
\title{Summarizes the Estimated Parameters of Fitted glca Model}
\usage{
\method{summary}{glca}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of "\code{glca}", usually, a result of a call to \code{glca}}

\item{digits}{the number of digits to be printed}

\item{...}{further arguments passed to or from other methods}
}
\value{
This function prints decriptions of model and its more detailed estimated parameters but returns \code{NULL}.
}
\description{
\code{summary} method for class "\code{glca}".
}
\examples{
## For examples see example(glca)

}
\seealso{
\code{\link{glca}}
}
